+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.23 (pre-release):
 * [Change] `KSP 1.8` compatibility. __WARNING__: the mod won't work with version lower than `KSP 1.8`!
 * [Change] Replace `Editor/partGrabModifiers` setting with `Editor/editorPartGrabAction`.
+* [Enhancement] Allow dropping KIS items on the kerbal: [Dragging equipped items](https://youtu.be/udTj7_pO2hc). Thanks to [sleepyquelea](https://github.com/sleepyquelea)!
 * [Fix #339] Inflatable habitat parts have unreasonable inventories.
 * [Fix #341] AddPodInventories not called for initial crew capacity 0.